Improve specialisation for imported functions
At a SPECIALSE pragma for an imported Id, we used to check that
it was marked INLINABLE.  But that turns out to interact badly with
worker/wrapper: see Note [Worker-wrapper for INLINABLE functions] in
GHC.Core.Opt.WorkWrap.

So this small patch instead simply tests that we have an unfolding
for the function; see Note [SPECIALISE pragmas for imported Ids]
in GHC.Tc.Gen.Sig.

Fixes #19246