CPR analysis: Use CPR of scrutinee for Case Binder CPR (#19232)
For years we have lived in a rather sweet spot that gave case binders
the CPR property, unconditionally. Which is an optimistic hack that
is described in `Note [CPR in a DataAlt case alternative]`.

But with #18907, we are a bit less aggressive with unboxing arguments,
so we also have to be a bit less aggressive with giving case binders
the CPR property in order not to risk reboxing.

Fixes #19232.